backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
langchain
langchain-openai
langchain-community
faiss-cpu
scikit-learn
bs4
requests
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
pandas
